+++ src/main/java/jp/main/taikun/mysticaleverything/ItemCompressionRecipe.java
@@ -22,10 +22,12 @@ public class ItemCompressionRecipe extends CustomRecipe {
         if (catalyst.is(Mysticaleverything.COMPRESSION_CATALYST.get())) {
             if (TagItemHelper.hasResource(catalyst)) return false;
             ItemStack first = container.getItem(0);
+            if (!Config.filter(first)) return false;
+            boolean ignoreNBT = Config.disableNBT(first);
             for (int i = 1; i < 9; i++){
                 if (i == 4) continue;
                 ItemStack current = container.getItem(i);
-                if (Config.DISABLE_NBT.get()){
+                if (ignoreNBT){
                     if (!current.is(first.getItem())) {
                         return false;
                     }
@@ -43,7 +45,7 @@ public class ItemCompressionRecipe extends CustomRecipe {
     @Override
     public @NotNull ItemStack assemble(CraftingContainer container, @NotNull RegistryAccess access) {
         ItemStack first = container.getItem(0);
-        if (Config.DISABLE_NBT.get()) {
+        if (Config.disableNBT(first)) {
             first = first.copy();
             first.setTag(null);
         }