+++ src/main/java/jp/main/taikun/mysticaleverything/mixin/main/InfusionAlterTileEntityMixin.java
@@ -49,7 +49,7 @@ public class InfusionAlterTileEntityMixin {
             }
             ItemStack outputItem = Mysticaleverything.EVERYTHING_CROP.get().asItem().getDefaultInstance();
             outputItem.setCount(1);
-            if (Config.DISABLE_NBT.get()) {
+            if (Config.disableNBT(ing)) {
                 ing = new ItemStack(ing.getItem(), 1);
                 TagItemHelper.setResource(outputItem, ing);
             } else {